Added some randomization and upgrade buttons and functionality
Tried to make it so cars are able to hit the checkpoints easier without all following each other. On a large scale, it seemed to work a little bit.

Added upgrades to buy more cars and enhance passive currency generation. Mainly placeholders or rudimentary design as I work on it.